requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
streamlit==1.32.0
pydantic==2.6.4
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
fpdf2==2.7.9
validators==0.22.0